move governance msg and state
