+++ contracts/governance/examples/schema.rs
@@ -3,8 +3,8 @@ use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use governance::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
-use governance::state::State;
+use white_whale::governance::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use white_whale::governance::state::State;
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
