+++ contracts/governance/src/lib.rs
@@ -2,9 +2,7 @@ pub use crate::error::ContractError;
 
 pub mod contract;
 mod error;
-pub mod msg;
 mod staking;
-pub mod state;
 
 #[cfg(not(target_arch = "wasm32"))]
 #[cfg(test)]
