+++ contracts/governance/src/staking.rs
@@ -1,10 +1,10 @@
 use crate::error::ContractError;
-use crate::state::{
+use white_whale::governance::state::{
     bank_read, bank_store, config_read, config_store, poll_read, poll_voter_store, state_read,
     state_store, Config, Poll, State, TokenManager,
 };
 
-use crate::state::{PollStatus, StakerResponse};
+use white_whale::governance::state::{PollStatus, StakerResponse};
 use cosmwasm_std::{
     to_binary, Addr, CanonicalAddr, CosmosMsg, Deps, DepsMut, MessageInfo, Response, StdResult,
     Storage, Uint128, WasmMsg,
