+++ contracts/governance/src/tests/instantiate.rs
@@ -3,8 +3,8 @@ use cosmwasm_std::{coins, from_binary, DepsMut};
 use cosmwasm_std::{Api, CanonicalAddr, Decimal, Uint128};
 
 use crate::contract::{execute, instantiate, query};
-use crate::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
-use crate::state::{config_read, state_read, Config, ConfigResponse, State};
+use white_whale::governance::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use white_whale::governance::state::{config_read, state_read, Config, ConfigResponse, State};
 use crate::tests::common::{
     DEFAULT_EXPIRATION_PERIOD, DEFAULT_FIX_PERIOD, DEFAULT_PROPOSAL_DEPOSIT, DEFAULT_QUORUM,
     DEFAULT_THRESHOLD, DEFAULT_TIMELOCK_PERIOD, DEFAULT_VOTING_PERIOD, TEST_CREATOR, VOTING_TOKEN,
