+++ contracts/governance/src/tests/integration_test.rs
@@ -4,9 +4,9 @@ use cosmwasm_std::testing::{mock_env, MockApi, MockStorage};
 use cosmwasm_std::{to_binary, Addr, BlockInfo, Decimal, Empty, Timestamp, Uint128};
 use cw_multi_test::{App, BankKeeper, Contract, ContractWrapper, Executor};
 
-use crate::msg::{ExecuteMsg, InstantiateMsg};
+use white_whale::governance::msg::{ExecuteMsg, InstantiateMsg};
 
-use crate::state::{Cw20HookMsg, PollExecuteMsg, VoteOption};
+use white_whale::governance::state::{Cw20HookMsg, PollExecuteMsg, VoteOption};
 use crate::tests::common::{
     DEFAULT_EXPIRATION_PERIOD, DEFAULT_FIX_PERIOD, DEFAULT_QUORUM, DEFAULT_THRESHOLD,
     DEFAULT_TIMELOCK_PERIOD, DEFAULT_VOTING_PERIOD,
