+++ contracts/governance/src/tests/poll.rs
@@ -7,8 +7,8 @@ use cw20::{Cw20ExecuteMsg, Cw20ReceiveMsg};
 use terraswap::querier::query_token_balance;
 
 use crate::contract::{execute, query};
-use crate::msg::{ExecuteMsg, QueryMsg};
-use crate::state::{
+use white_whale::governance::msg::{ExecuteMsg, QueryMsg};
+use white_whale::governance::state::{
     bank_read, poll_voter_read, state_read, Cw20HookMsg, OrderBy, PollExecuteMsg, PollResponse,
     PollStatus, PollsResponse, StakerResponse, State, VoteOption, VoterInfo, VotersResponse,
 };
