+++ contracts/governance/src/tests/staking.rs
@@ -3,9 +3,9 @@ use cosmwasm_std::{attr, from_binary, to_binary, Addr, Uint128};
 use cw20::Cw20ReceiveMsg;
 
 use crate::contract::{execute, query};
-use crate::msg::{ExecuteMsg, QueryMsg};
+use white_whale::governance::msg::{ExecuteMsg, QueryMsg};
 use crate::staking::stake_voting_tokens;
-use crate::state::{Cw20HookMsg, StakerResponse};
+use white_whale::governance::state::{Cw20HookMsg, StakerResponse};
 use crate::tests::common::{TEST_VOTER, VOTING_TOKEN};
 use crate::tests::mock_querier::mock_dependencies;
 use crate::tests::{instantiate, poll};
