+++ packages/white_whale/src/governance/mod.rs
@@ -1,0 +1,2 @@
+pub mod msg; 
+pub mod state;
