+++ packages/white_whale/src/lib.rs
@@ -14,3 +14,4 @@ pub mod tokenomics;
 pub mod treasury;
 pub mod ust_vault;
 pub mod vesting;
+pub mod governance;
